+++ src/components/App/People/PeopleList.js
@@ -8,12 +8,11 @@ import { BasicActionBtn } from "../../../~reusables/atoms/Buttons";
 // styles
 import { heading_5, body_2 } from "../../../~reusables/variables/font-sizes";
 import {
-  support,
-  light_support,
-  primary
+  support
 } from "../../../~reusables/variables/colors";
 import { xs_space } from "../../../~reusables/variables/spacing";
 import { tablet_max_width } from "../../../~reusables/variables/media-queries";
+import Tabs from "../../../~reusables/molecules/Tabs";
 
 const PeopleList = props => {
   const { firstHeading, secondHeading, thirdHeading, listOfPeople } = props;
@@ -23,39 +22,8 @@ const PeopleList = props => {
     setTabIndex(0);
   }, [listOfPeople]);
 
-  let tabs = 1;
-  if (listOfPeople.length > 8) {
-    tabs = Math.ceil(listOfPeople.length / 8);
-  }
-
-  let leftGreyedOut;
-  if (tabIndex < 1) {
-    leftGreyedOut = "left-greyed-out";
-  }
-
-  let rightGreyedOut;
-  if (tabIndex === tabs - 1) {
-    rightGreyedOut = "right-greyed-out";
-  }
-
-  const renderTabs = () => {
-    let tabsDisplay = [];
-    for (let i = 0; i < tabs; i++) {
-      tabsDisplay.push(
-        <BasicActionBtn
-          onClick={e => setTabIndex(i)}
-          className={tabIndex === i ? "active-button" : ""}
-          key={i + 1}
-        >
-          {i + 1}
-        </BasicActionBtn>
-      );
-    }
-    return tabsDisplay;
-  };
-
   const renderPeople = () => {
-    return listOfPeople.map(person => {
+    return listOfPeople.slice(tabIndex * 8, 8 * (tabIndex + 1)).map(person => {
       return (
         <div key={person.name} className="body">
           <p>{person.name}</p>
@@ -79,15 +47,11 @@ const PeopleList = props => {
       </div>
       {renderPeople()}
       <div className="footer">
-        <p className={`${leftGreyedOut}`}>
-          <i className={`material-icons`}>keyboard_arrow_left</i>{" "}
-          <span>Back</span>
-        </p>
-        {renderTabs()}
-        <p className={`${rightGreyedOut}`} onClick={() => setTabIndex(1)}>
-          <span>Next</span>{" "}
-          <i className={`material-icons`}>keyboard_arrow_right</i>
-        </p>
+        <Tabs
+          tabIndex={tabIndex}
+          setTabIndex={setTabIndex}
+          arrayList={listOfPeople}
+        />
       </div>
     </StyledList>
   );
@@ -106,40 +70,6 @@ const StyledList = styled.div`
     }
     div {
       flex: 1 1 100px;
-    }
-  }
-
-  .footer {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    p {
-      color: ${primary};
-      font-weight: 500;
-      cursor: pointer;
-      display: flex;
-      align-items: center;
-    }
-    p:first-child {
-      margin-right: ${xs_space};
-    }
-
-    .left-greyed-out,
-    .right-greyed-out {
-      color: ${light_support};
-      cursor: not-allowed;
-    }
-
-    p:last-child {
-      margin-left: ${xs_space};
-    }
-
-    button {
-      margin: 0 ${xs_space};
-    }
-
-    .active-button {
-      background-color: ${primary};
     }
   }
